add opentitan AES simulation to gateware blocks
aes primitives are snapshotted into the simulation for now,
but might be interesting to consider turning this into a
submodule so we don't have to resync the files, except
submodules are messy and maybe in the end it'd be faster
just to manually copy the files in and check them in.

When automation takes longer than doing it by hand, is
it really automation?