modify the sha512 block behavior to work with `digest` API
The OpenTitan implementation is quite eager about finalizing
the data. If there is too long of a pause the engine automatically
finalizes the data (including padding etc) which is not really
what we want to have happen in a software-driven implementation.

The changes to the block ensure that data can only be finalized
when explicitly requested.

Also test infrastructure updated to use the Rust `digest` APIs.