engine25519 now passing add test + timing during p&r
the RF now is double-pumped, which allows me a bit more
control over timing. the downside is timing constraints are
a nightmare, but as long as you don't go and refactor the names
of signals willy-nilly everything will be okay.

i mean, really, who goes and just changes the names of signals
in someone's code just to make it more readable and/or complaint
with their coding standards...

oh wait.

good luck to that guy!