greatly improve timing closure, correct timing bug
Design had an extra register stage in the compute path which
wasn't supposed to be there; this caused it to fail tests
on real hardware because in fact we were trying to latch
results by far too early.

Removing that makes the loop from the RF-exec unit-RF fully
combinational (as intended, so that there is maximum time in the
eng_clk cycle to perform the computation). This created a bunch of new
multicycle paths that had to be annotated.