fix race condition in WDT reset with a double-trigger interlock
this patch means you have to "miss twice in a row" before a reset
happens. There is a race condition where you may have disarmed the
WDT, but then it just so happened to pop back into an armed state
just before the reset happens, which will cause spurious resets.

In order to prevent this, it falls back into a non-resetting mode,
and only if the WDT interval passes twice without progressing to
the disarmed state will it reset the CPU.

Again, this dance is because the WDT is asynchronous *and variable*
to the CPU (because it's a ring oscillator), so we have all the
clock domain crossing issues.