remove jtag_phy from the CI set
Rust broke allocators; the failure in CI is due to a Rust change.
However, we're about to introduce this module into the Xous
code base as native code, so let's wait until then instead of
recoding the module specifically to patch around this.

This is a relatively low-risk module, anyways.