make the result of the force_round appear on the output of chacha
previously, the extra mixing did nothing because it just modified
internal state but failed to update the output. Now, the output
gets update (in a pipelined fashion).

Simulations show that if you're reading at maximal rates, you
will only get the minimum guaranteed number of rounds (which is
still in the recommended amounts for Chacha) but if you have
some idle time between reads (which is very typical) you will
get some variable extra number of rounds, which was the
goal of the force_round function.