relax INMODE critical path on engine
simplify the state decode for the INMODE mux on the
DSP blocks; previously, all of the current master state machine
plus multiply stepping had to be considered on each cycle.

The new system reduces the INMODE to just four states (2 bits)
that have to be tracked, plus the current multiply step. Results
show this path has been completely eliminated from the critical
path in slack histograms, and it still passes simulation, so
I think we're good.