fix repeated key_ack when pending is not immediately serviced
pulse synchronizer will keep on pulsing the output until the pending
bit is cleared, which was wreaking havoc with the state of the keyboard
machine.

test bench suite was also improved to model the test case that
causes this issue (namely, the case that someone presses the keyboard
during the resume process while Xous is not yet running, and
causes a pending, which is not immediately acknowledged, leading
to the pulse train that was fixed by this bug).