remove the default_nets directive
vivado has a quirk in that this directive sticks beyond the scope
of the verilog module, so modules added later in the build order
inherit this directive.

SpinalHDL generated code assumes a default net of `wire`. Thus,
when these two types are in the same project, there is an inherent
ordering required of when the modules are created.

While I like the disicpline of the directive, it's not uniformly
applied and so I think the best choice is to remove the directive.