add panic handler to the basic bios library, remove the proc_macro
the proc_macro is now broken in rust 1.70 because for some reason
when you try to compile a proc macro it includes both the
target and the source object files into the macro and the linker
gets confused because it has two different binary formats.

this patch set works around that.