pin svd2rust version
version 0.31.0 broke a whole bunch of APIs by refactoring to require
a method call instead of a direct access to get to peripherals.

The error looks a bit like this:

```
00:09:40.188 error[E0616]: field `digest7` of struct `sha2_pac::sha2::RegisterBlock` is private
00:09:40.188    --> sim/sha2/testbench/sha2-hal/src/hal_sha2.rs:120:48
00:09:40.188     |
00:09:40.188 120 |                 7 => digest[reg] = self.p.SHA2.digest7.read().bits(),
00:09:40.188     |                                                ^^^^^^^ private field
00:09:40.188     |
00:09:40.188 help: a method `digest7` also exists, call it with parentheses
00:09:40.188     |
00:09:40.188 120 |                 7 => digest[reg] = self.p.SHA2.digest7().read().bits(),
00:09:40.188     |                                                       ++
```

So previously you could just do self.p.SHA2.digest7 to access
things, now you have to do self.p.SHA2.digest7(). Not really worth it
I think at this point to refactor dozens of tests to follow the svd2rust
crate since it's a code gen for accessors to peripherals: this kind of
stuff should ideally be immutable anyways.